requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
psycopg2-binary==2.9.7
neo4j==5.13.0
redis==5.0.1
python-dotenv==1.0.0
Pillow==10.0.1
python-magic==0.4.27
werkzeug==2.3.7
marshmallow==3.20.1
email-validator==2.0.0
bcrypt==4.0.1
requests==2.31.0

# 测试相关
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0

# 开发工具
black==23.9.1
flake8==6.1.0
flasgger==0.9.7.1
